requirements.txt
discord.py==2.3.2
motor==3.3.1
python-dotenv==1.0.1
